Initial refactor to group modifiers into separate modules
