+++ satpy/tests/test_modifiers.py
@@ -62,7 +62,7 @@ class TestSunZenithCorrector(unittest.TestCase):
 
     def test_basic_default_not_provided(self):
         """Test default limits when SZA isn't provided."""
-        from satpy.modifiers import SunZenithCorrector
+        from satpy.modifiers.geometry import SunZenithCorrector
         comp = SunZenithCorrector(name='sza_test', modifiers=tuple())
         res = comp((self.ds1,), test_attr='test')
         np.testing.assert_allclose(res.values, np.array([[22.401667, 22.31777], [22.437503, 22.353533]]))
@@ -76,21 +76,21 @@ class TestSunZenithCorrector(unittest.TestCase):
 
     def test_basic_lims_not_provided(self):
         """Test custom limits when SZA isn't provided."""
-        from satpy.modifiers import SunZenithCorrector
+        from satpy.modifiers.geometry import SunZenithCorrector
         comp = SunZenithCorrector(name='sza_test', modifiers=tuple(), correction_limit=90)
         res = comp((self.ds1,), test_attr='test')
         np.testing.assert_allclose(res.values, np.array([[66.853262, 68.168939], [66.30742, 67.601493]]))
 
     def test_basic_default_provided(self):
         """Test default limits when SZA is provided."""
-        from satpy.modifiers import SunZenithCorrector
+        from satpy.modifiers.geometry import SunZenithCorrector
         comp = SunZenithCorrector(name='sza_test', modifiers=tuple())
         res = comp((self.ds1, self.sza), test_attr='test')
         np.testing.assert_allclose(res.values, np.array([[22.401667, 22.31777], [22.437503, 22.353533]]))
 
     def test_basic_lims_provided(self):
         """Test custom limits when SZA is provided."""
-        from satpy.modifiers import SunZenithCorrector
+        from satpy.modifiers.geometry import SunZenithCorrector
         comp = SunZenithCorrector(name='sza_test', modifiers=tuple(), correction_limit=90)
         res = comp((self.ds1, self.sza), test_attr='test')
         np.testing.assert_allclose(res.values, np.array([[66.853262, 68.168939], [66.30742, 67.601493]]))
@@ -98,7 +98,7 @@ class TestSunZenithCorrector(unittest.TestCase):
     def test_imcompatible_areas(self):
         """Test sunz correction on incompatible areas."""
         from satpy.composites import IncompatibleAreas
-        from satpy.modifiers import SunZenithCorrector
+        from satpy.modifiers.geometry import SunZenithCorrector
         comp = SunZenithCorrector(name='sza_test', modifiers=tuple(), correction_limit=90)
         with pytest.raises(IncompatibleAreas):
             comp((self.ds2, self.sza), test_attr='test')
@@ -148,15 +148,15 @@ class TestNIRReflectance(unittest.TestCase):
             return self.refl_with_co2
         return self.refl
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_provide_sunz_no_co2(self, calculator, apply_modifier_info, sza):
         """Test NIR reflectance compositor provided only sunz."""
         calculator.return_value = mock.MagicMock(
             reflectance_from_tbs=self.refl_from_tbs)
         sza.return_value = self.da_sunz
-        from satpy.modifiers import NIRReflectance
+        from satpy.modifiers.spectral import NIRReflectance
 
         comp = NIRReflectance(name='test')
         info = {'modifiers': None}
@@ -167,15 +167,15 @@ class TestNIRReflectance(unittest.TestCase):
         assert res.attrs['sun_zenith_threshold'] is not None
         assert np.allclose(res.data, self.refl * 100).compute()
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_no_sunz_no_co2(self, calculator, apply_modifier_info, sza):
         """Test NIR reflectance compositor with minimal parameters."""
         calculator.return_value = mock.MagicMock(
             reflectance_from_tbs=self.refl_from_tbs)
         sza.return_value = self.da_sunz
-        from satpy.modifiers import NIRReflectance
+        from satpy.modifiers.spectral import NIRReflectance
 
         comp = NIRReflectance(name='test')
         info = {'modifiers': None}
@@ -186,14 +186,14 @@ class TestNIRReflectance(unittest.TestCase):
         self.refl_from_tbs.assert_called_with(self.da_sunz, self.nir.data, self.ir_.data, tb_ir_co2=None)
         assert np.allclose(res.data, self.refl * 100).compute()
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_no_sunz_with_co2(self, calculator, apply_modifier_info, sza):
         """Test NIR reflectance compositor provided extra co2 info."""
         calculator.return_value = mock.MagicMock(
             reflectance_from_tbs=self.refl_from_tbs)
-        from satpy.modifiers import NIRReflectance
+        from satpy.modifiers.spectral import NIRReflectance
         sza.return_value = self.da_sunz
 
         comp = NIRReflectance(name='test')
@@ -207,14 +207,14 @@ class TestNIRReflectance(unittest.TestCase):
         self.refl_from_tbs.assert_called_with(self.da_sunz, self.nir.data, self.ir_.data, tb_ir_co2=co2.data)
         assert np.allclose(res.data, self.refl_with_co2 * 100).compute()
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_provide_sunz_and_threshold(self, calculator, apply_modifier_info, sza):
         """Test NIR reflectance compositor provided sunz and a sunz threshold."""
         calculator.return_value = mock.MagicMock(
             reflectance_from_tbs=self.refl_from_tbs)
-        from satpy.modifiers import NIRReflectance
+        from satpy.modifiers.spectral import NIRReflectance
         sza.return_value = self.da_sunz
 
         comp = NIRReflectance(name='test', sunz_threshold=84.0)
@@ -224,12 +224,12 @@ class TestNIRReflectance(unittest.TestCase):
         self.assertEqual(res.attrs['sun_zenith_threshold'], 84.0)
         calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039', sunz_threshold=84.0)
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_sunz_threshold_default_value_is_not_none(self, calculator, apply_modifier_info, sza):
         """Check that sun_zenith_threshold is not None."""
-        from satpy.modifiers import NIRReflectance
+        from satpy.modifiers.spectral import NIRReflectance
 
         comp = NIRReflectance(name='test', sunz_threshold=None)
         info = {'modifiers': None}
@@ -243,9 +243,9 @@ class TestNIRReflectance(unittest.TestCase):
 class TestNIREmissivePartFromReflectance(unittest.TestCase):
     """Test the NIR Emissive part from reflectance compositor."""
 
-    @mock.patch('satpy.modifiers.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.modifiers.Calculator')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
     def test_compositor(self, calculator, apply_modifier_info, sza):
         """Test the NIR emissive part from reflectance compositor."""
         import numpy as np
@@ -265,7 +265,7 @@ class TestNIREmissivePartFromReflectance(unittest.TestCase):
         emissive_part.return_value = emissive
         calculator.return_value = mock.MagicMock(emissive_part_3x=emissive_part)
 
-        from satpy.modifiers import NIREmissivePartFromReflectance
+        from satpy.modifiers.spectral import NIREmissivePartFromReflectance
 
         comp = NIREmissivePartFromReflectance(name='test', sunz_threshold=86.0)
         info = {'modifiers': None}
@@ -323,7 +323,7 @@ class TestPSPAtmosphericalCorrection(unittest.TestCase):
         self.module_patcher.stop()
 
     @mock.patch('satpy.modifiers.PSPAtmosphericalCorrection.apply_modifier_info')
-    @mock.patch('satpy.modifiers.get_satpos')
+    @mock.patch('satpy.modifiers.atmosphere.get_satpos')
     def test_call(self, get_satpos, *mocks):
         """Test atmospherical correction."""
         from satpy.modifiers import PSPAtmosphericalCorrection
@@ -367,7 +367,7 @@ class TestPSPRayleighReflectance(unittest.TestCase):
         """Unpatch in-class imports."""
         self.module_patcher.stop()
 
-    @mock.patch('satpy.modifiers.get_satpos')
+    @mock.patch('satpy.modifiers.atmosphere.get_satpos')
     def test_get_angles(self, get_satpos):
         """Test sun and satellite angle calculation."""
         from satpy.modifiers import PSPRayleighReflectance
